requirements.txt
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
langserve
langchain-google-genai
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15

aiohttp
diskcache
fpdf
openpyxl
